--- src/redux/missons/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import missionReducer from './missionReducer';
-
-const reducers = combineReducers({
-  allMissions: missionReducer,
-});
-
-export default reducers;
