+++ src/redux/missons/store.js
@@ -1,7 +1,16 @@
-import { configureStore, applyMiddleware } from '@reduxjs/toolkit';
+import { configureStore } from '@reduxjs/toolkit';
+import { combineReducers, applyMiddleware } from 'redux';
+import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import reducers from './reducers/index';
+import missionReducer from './reducers/missionReducer';
 
-const store = configureStore(reducers, applyMiddleware(thunk));
+const reducers = combineReducers({
+  allMissions: missionReducer,
+});
+
+const store = configureStore(
+  { reducer: reducers },
+  applyMiddleware(logger, thunk),
+);
 
 export default store;
